requirements.txt
openai>=1.2.0
anthropic>=0.6.8
pydantic>=2.6.3
langgraph>=0.0.20
streamlit>=1.32.0
selenium>=4.19.0
requests>=2.31.0
google-api-python-client>=2.108.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
